# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Instantly Service",
    "description": "Cold email outreach service via Instantly.ai API. Manages campaigns, leads, accounts, analytics, and webhook events.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3011"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CreateCampaignResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "campaign": {
            "$ref": "#/components/schemas/CampaignSummary"
          }
        },
        "required": [
          "success",
          "campaign"
        ]
      },
      "CampaignSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "instantlyCampaignId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "instantlyCampaignId",
          "name",
          "status"
        ]
      },
      "CreateCampaignRequest": {
        "type": "object",
        "properties": {
          "brandIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Brand IDs associated with this campaign"
          },
          "workflowSlug": {
            "type": "string",
            "description": "Workflow slug for stats grouping"
          },
          "name": {
            "type": "string"
          },
          "accountIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          }
        },
        "required": [
          "brandIds",
          "name"
        ]
      },
      "UpdateStatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "completed"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "CheckStatusResponse": {
        "type": "object",
        "properties": {
          "checked": {
            "type": "number",
            "description": "Number of active campaigns checked"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "instantlyCampaignId": {
                  "type": "string"
                },
                "campaignId": {
                  "type": "string",
                  "nullable": true
                },
                "leadEmail": {
                  "type": "string",
                  "nullable": true
                },
                "reason": {
                  "type": "string"
                }
              },
              "required": [
                "instantlyCampaignId",
                "campaignId",
                "leadEmail",
                "reason"
              ]
            },
            "description": "Campaigns that entered error state"
          }
        },
        "required": [
          "checked",
          "errors"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/campaigns": {
      "post": {
        "summary": "Create a campaign",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCampaignResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to create campaign",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/campaigns/{campaignId}": {
      "get": {
        "summary": "Get a campaign",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "campaignId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign found"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Campaign not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/campaigns/by-org/{orgId}": {
      "get": {
        "summary": "List campaigns by organization",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Campaigns list"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/campaigns/{campaignId}/status": {
      "patch": {
        "summary": "Update campaign status",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "campaignId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          },
          "400": {
            "description": "Invalid status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Campaign not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/campaigns/check-status": {
      "post": {
        "summary": "Poll active campaigns for errors",
        "description": "Checks all active campaigns against the Instantly API to detect error states. For each errored campaign: updates DB status, cancels provisioned costs, fails the run, and sends an admin notification.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Status check complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}